Invoke tickle rather than tickle.sendTransaction
In the truffle [Executing Contract Functions](https://truffle.readthedocs.io/en/develop/getting_started/contracts/#executing-contract-functions)
and it explains in the Abstraction section that contract methods by
default will use sendTransactions thus using .sendTransaction is
superfluous and this change removes it.

But in the subsection "Making a Call" that .call should be used to
invoke methods that don't change the contract.